Make sure not to break clipboard chain when exiting
By not stopping clipboard monitoring, the handler remains in the clipboard chain even after the program exists. Since it can not handle anything anymore, the clipboard chain is broken and subsequent monitors (such as other instance of PasteIntoFile) will no longer receive any updates when something is copied to the clipboard.

Previously, this was breaking the batch mode as well as clipboard patching mechanism when the user opened and closed a second instance, e.g. via the context menu.